chore: move wait utilities to utils directory
