+++ webui/react/src/pages/TrialDetails/TrialDetailsHeader.tsx
@@ -11,8 +11,8 @@ import { openOrCreateTensorBoard } from 'services/api';
 import Icon from 'shared/components/Icon/Icon';
 import { ExperimentAction as Action, ExperimentAction, ExperimentBase, TrialDetails } from 'types';
 import { canUserActionExperiment } from 'utils/experiment';
+import { openCommand } from 'utils/wait';
 import { getWorkload, isMetricsWorkload } from 'utils/workload';
-import { openCommand } from 'wait';
 
 export const trialWillNeverHaveData = (trial: TrialDetails): boolean => {
   const isTerminal = terminalRunStates.has(trial.state);
