+++ webui/react/src/pages/Wait.tsx
@@ -13,7 +13,7 @@ import { ErrorType } from 'shared/utils/error';
 import { capitalize } from 'shared/utils/string';
 import { CommandState } from 'types';
 import handleError from 'utils/error';
-import { WaitStatus } from 'wait';
+import { WaitStatus } from 'utils/wait';
 
 import css from './Wait.module.scss';
